Split Resend node into descriptions and helpers
