+++ nodes/Resend/Resend.node.ts
@@ -1,13 +1,40 @@
 import {
 	IExecuteFunctions,
-	ILoadOptionsFunctions,
 	INodeExecutionData,
-	INodePropertyOptions,
 	INodeType,
 	INodeTypeDescription,
 	NodeOperationError,
 	NodeConnectionType,
 } from 'n8n-workflow';
+import {
+	apiKeyFields,
+	apiKeyOperations,
+	broadcastFields,
+	broadcastOperations,
+	contactFields,
+	contactOperations,
+	domainFields,
+	domainOperations,
+	emailFields,
+	emailOperations,
+	segmentFields,
+	segmentOperations,
+	templateFields,
+	templateOperations,
+	topicFields,
+	topicOperations,
+} from './descriptions';
+import {
+	buildTemplateSendVariables,
+	getSegments,
+	getTemplateVariables,
+	getTemplates,
+	getTopics,
+	ListOptions,
+	normalizeEmailList,
+	parseTemplateVariables,
+	requestList,
+} from './GenericFunctions';
 
 export class Resend implements INodeType {
 	description: INodeTypeDescription = {
@@ -79,2843 +106,37 @@ export class Resend implements INodeType {
 				default: 'email',
 			},
 
-			// EMAIL OPERATIONS
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: {
-					show: {
-						resource: ['email'],
-					},
-				}, options: [
-					{
-						name: 'Cancel',
-						value: 'cancel',
-						description: 'Cancel a scheduled email',
-						action: 'Cancel an email',
-					},
-					{
-						name: 'List',
-						value: 'list',
-						description: 'List sent emails',
-						action: 'List emails',
-					},
-					{
-						name: 'Retrieve',
-						value: 'retrieve',
-						description: 'Retrieve an email by ID',
-						action: 'Retrieve an email',
-					},
-					{
-						name: 'Send',
-						value: 'send',
-						description: 'Send an email',
-						action: 'Send an email',
-					},
-					{
-						name: 'Send Batch',
-						value: 'sendBatch',
-						description: 'Send multiple emails at once',
-						action: 'Send batch emails',
-					},
-					{
-						name: 'Update',
-						value: 'update',
-						description: 'Update an email',
-						action: 'Update an email',
-					},
-				],
-				default: 'send',
-			},
-			// TEMPLATE OPERATIONS
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: {
-					show: {
-						resource: ['templates'],
-					},
-				},
-				options: [
-					{
-						name: 'Create',
-						value: 'create',
-						description: 'Create a new template',
-						action: 'Create a template',
-					},
-					{
-						name: 'Delete',
-						value: 'delete',
-						description: 'Delete a template',
-						action: 'Delete a template',
-					},
-					{
-						name: 'Get',
-						value: 'get',
-						description: 'Get a template by ID',
-						action: 'Get a template',
-					},
-					{
-						name: 'List',
-						value: 'list',
-						description: 'List all templates',
-						action: 'List templates',
-					},
-					{
-						name: 'Update',
-						value: 'update',
-						description: 'Update a template',
-						action: 'Update a template',
-					},
-				],
-				default: 'list',
-			},
-			// DOMAIN OPERATIONS
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: {
-					show: {
-						resource: ['domains'],
-					},
-				},
-				options: [
-					{
-						name: 'Create',
-						value: 'create',
-						description: 'Create a new domain',
-						action: 'Create a domain',
-					},
-					{
-						name: 'Delete',
-						value: 'delete',
-						description: 'Delete a domain',
-						action: 'Delete a domain',
-					},
-					{
-						name: 'Get',
-						value: 'get',
-						description: 'Get a domain by ID',
-						action: 'Get a domain',
-					},
-					{
-						name: 'List',
-						value: 'list',
-						description: 'List all domains',
-						action: 'List domains',
-					},
-					{
-						name: 'Update',
-						value: 'update',
-						description: 'Update a domain',
-						action: 'Update a domain',
-					},
-					{
-						name: 'Verify',
-						value: 'verify',
-						description: 'Verify a domain',
-						action: 'Verify a domain',
-					},
-				],
-				default: 'list',
-			},
-			// API KEY OPERATIONS
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: {
-					show: {
-						resource: ['apiKeys'],
-					},
-				},
-				options: [
-					{
-						name: 'Create',
-						value: 'create',
-						description: 'Create a new API key',
-						action: 'Create an API key',
-					},
-					{
-						name: 'Delete',
-						value: 'delete',
-						description: 'Delete an API key',
-						action: 'Delete an API key',
-					},
-					{
-						name: 'List',
-						value: 'list',
-						description: 'List all API keys',
-						action: 'List API keys',
-					},
-				],
-				default: 'list',
-			},
-			// BROADCAST OPERATIONS
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-					},
-				},
-				options: [
-					{
-						name: 'Create',
-						value: 'create',
-						description: 'Create a new broadcast',
-						action: 'Create a broadcast',
-					},
-					{
-						name: 'Delete',
-						value: 'delete',
-						description: 'Delete a broadcast',
-						action: 'Delete a broadcast',
-					},
-					{
-						name: 'Get',
-						value: 'get',
-						description: 'Get a broadcast by ID',
-						action: 'Get a broadcast',
-					},
-					{
-						name: 'List',
-						value: 'list',
-						description: 'List all broadcasts',
-						action: 'List broadcasts',
-					},
-					{
-						name: 'Send',
-						value: 'send',
-						description: 'Send a broadcast',
-						action: 'Send a broadcast',
-					},
-					{
-						name: 'Update',
-						value: 'update',
-						description: 'Update a broadcast',
-						action: 'Update a broadcast',
-					},
-				],
-				default: 'list',
-			},
-			// SEGMENT OPERATIONS
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: {
-					show: {
-						resource: ['segments'],
-					},
-				},
-				options: [
-					{
-						name: 'Create',
-						value: 'create',
-						description: 'Create a new segment',
-						action: 'Create a segment',
-					},
-					{
-						name: 'Get',
-						value: 'get',
-						description: 'Get a segment by ID',
-						action: 'Get a segment',
-					},
-					{
-						name: 'Delete',
-						value: 'delete',
-						description: 'Delete a segment',
-						action: 'Delete a segment',
-					},
-					{
-						name: 'List',
-						value: 'list',
-						description: 'List all segments',
-						action: 'List segments',
-					},
-				],
-				default: 'list',
-			},
-			// TOPIC OPERATIONS
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: {
-					show: {
-						resource: ['topics'],
-					},
-				},
-				options: [
-					{
-						name: 'Create',
-						value: 'create',
-						description: 'Create a new topic',
-						action: 'Create a topic',
-					},
-					{
-						name: 'Delete',
-						value: 'delete',
-						description: 'Delete a topic',
-						action: 'Delete a topic',
-					},
-					{
-						name: 'Get',
-						value: 'get',
-						description: 'Get a topic by ID',
-						action: 'Get a topic',
-					},
-					{
-						name: 'List',
-						value: 'list',
-						description: 'List all topics',
-						action: 'List topics',
-					},
-					{
-						name: 'Update',
-						value: 'update',
-						description: 'Update a topic',
-						action: 'Update a topic',
-					},
-				],
-				default: 'list',
-			},
-			// CONTACT OPERATIONS
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: {
-					show: {
-						resource: ['contacts'],
-					},
-				},
-				options: [
-					{
-						name: 'Create',
-						value: 'create',
-						description: 'Create a new contact',
-						action: 'Create a contact',
-					},
-					{
-						name: 'Delete',
-						value: 'delete',
-						description: 'Delete a contact',
-						action: 'Delete a contact',
-					},
-					{
-						name: 'Get',
-						value: 'get',
-						description: 'Get a contact by ID',
-						action: 'Get a contact',
-					},
-					{
-						name: 'List',
-						value: 'list',
-						description: 'List contacts',
-						action: 'List contacts',
-					},
-					{
-						name: 'Update',
-						value: 'update',
-						description: 'Update a contact',
-						action: 'Update a contact',
-					},
-				],
-				default: 'list',
-			},
-			// Properties for "Send Email" operation
-			{
-				displayName: 'From',
-				name: 'from',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'you@example.com',
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-					},
-				},
-				description: 'Sender email address. To include a friendly name, use the format "Your Name &lt;sender@domain.com&gt;".',
-			},
-			{
-				displayName: 'To',
-				name: 'to',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'user@example.com',
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-					},
-				},
-				description: 'Recipient email address. For multiple addresses, separate with commas (max 50).',
-			},
-			{
-				displayName: 'Subject',
-				name: 'subject',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'Hello from n8n!',
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-					},
-				},
-				description: 'Email subject line',
-			},
-			{
-				displayName: 'Use Template',
-				name: 'useTemplate',
-				type: 'boolean',
-				default: false,
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-					},
-				},
-				description: 'Whether to send using a published template instead of HTML/Text content',
-			},
-			{
-				displayName: 'Email Format',
-				name: 'emailFormat',
-				type: 'options',
-				options: [
-					{
-						name: 'HTML',
-						value: 'html',
-						description: 'Send email with HTML content',
-					},
-					{
-						name: 'HTML and Text',
-						value: 'both',
-						description: 'Send email with both HTML and text content',
-					},
-					{
-						name: 'Text',
-						value: 'text',
-						description: 'Send email with plain text content',
-					},
-				],
-				default: 'html',
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-						useTemplate: [false],
-					},
-				},
-				description: 'Choose the format for your email content. HTML allows rich formatting, text is simple and universally compatible.',
-			},
-			{
-				displayName: 'Template Name or ID',
-				name: 'emailTemplateId',
-				type: 'options',
-				required: true,
-				default: '',
-				placeholder: '34a080c9-b17d-4187-ad80-5af20266e535',
-				typeOptions: {
-					loadOptionsMethod: 'getTemplates',
-				},
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-						useTemplate: [true],
-					},
-				},
-				description: 'Select a template or enter an ID/alias using an expression. Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>.',
-			},
-			{
-				displayName: 'Template Variables',
-				name: 'emailTemplateVariables',
-				type: 'fixedCollection',
-				default: { variables: [] },
-				typeOptions: {
-					multipleValues: true,
-				},
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-						useTemplate: [true],
-					},
-				},
-				description: 'Variables to render the template with',
-				options: [
-					{
-						name: 'variables',
-						displayName: 'Variable',
-						values: [
-							{
-								displayName: 'Key',
-								name: 'key',
-								type: 'options',
-								required: true,
-								default: '',
-								typeOptions: {
-									loadOptionsMethod: 'getTemplateVariables',
-									loadOptionsDependsOn: ['emailTemplateId'],
-									allowCustomValues: true,
-								},
-								description: 'Template variable name',
-							},
-							{
-								displayName: 'Value',
-								name: 'value',
-								type: 'string',
-								default: '',
-								description: 'Value for the template variable',
-							},
-						],
-					},
-				],
-			},
-			{
-				displayName: 'HTML Content',
-				name: 'html',
-				type: 'string',
-				default: '',
-				typeOptions: {
-					multiline: true,
-					rows: 4,
-				},
-				placeholder: '<p>Your HTML content here</p>',
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-						emailFormat: ['html', 'both'],
-						useTemplate: [false],
-					},
-				},
-				description: 'HTML version of the email content',
-			},
-			{
-				displayName: 'Text Content',
-				name: 'text',
-				type: 'string',
-				default: '',
-				typeOptions: {
-					multiline: true,
-					rows: 4,
-				},
-				placeholder: 'Your plain text content here',
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-						emailFormat: ['text', 'both'],
-						useTemplate: [false],
-					},
-				},
-				description: 'Plain text version of the email content',
-			},
-			{
-				displayName: 'Additional Options',
-				name: 'additionalOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['send'],
-					},
-				},				options: [
-					{
-						displayName: 'Attachments',
-						name: 'attachments',
-						type: 'fixedCollection',
-						default: { attachments: [] },
-						typeOptions: {
-							multipleValues: true,
-						},
-						description: 'Email attachments (not supported with scheduled emails)',
-						options: [
-							{
-								name: 'attachments',
-								displayName: 'Attachment',
-								values: [
-									{
-										displayName: 'Attachment Type',
-										name: 'attachmentType',
-										type: 'options',
-										default: 'binaryData',
-										options: [
-											{
-												name: 'Binary Data',
-												value: 'binaryData',
-												description: 'Use binary data from previous node',
-											},
-											{
-												name: 'Remote URL',
-												value: 'url',
-												description: 'Use a URL to a remote file',
-											},
-										],
-									},
-									{
-										displayName: 'Binary Property',
-										name: 'binaryPropertyName',
-										type: 'string',
-										default: 'data',
-										placeholder: 'data',
-										description: 'Name of the binary property which contains the file data',
-										displayOptions: {
-											show: {
-												attachmentType: ['binaryData'],
-											},
-										},
-									},
-									{
-										displayName: 'Content ID',
-										name: 'content_id',
-										type: 'string',
-										default: '',
-										placeholder: 'image-1',
-										description: 'Content ID for embedding inline attachments via cid:',
-									},
-									{
-										displayName: 'Content Type',
-										name: 'content_type',
-										type: 'string',
-										default: '',
-										placeholder: 'image/png',
-										description: 'Content type for the attachment',
-									},
-									{
-										displayName: 'File Name',
-										name: 'filename',
-										type: 'string',
-										default: '',
-										placeholder: 'document.pdf',
-										description: 'Name for the attached file (required for both binary data and URL)',
-									},
-									{
-										displayName: 'File URL',
-										name: 'fileUrl',
-										type: 'string',
-										default: '',
-										placeholder: 'https://example.com/file.pdf',
-										description: 'URL to the remote file',
-										displayOptions: {
-											show: {
-												attachmentType: ['url'],
-											},
-										},
-									},
-								],
-							},
-						],
-					},
-					{
-						displayName: 'BCC',
-						name: 'bcc',
-						type: 'string',
-						default: '',
-						description: 'BCC recipient email addresses (comma-separated)',
-					},
-					{
-						displayName: 'CC',
-						name: 'cc',
-						type: 'string',
-						default: '',
-						description: 'CC recipient email addresses (comma-separated)',
-					},
-					{
-						displayName: 'Headers',
-						name: 'headers',
-						type: 'fixedCollection',
-						default: { headers: [] },
-						typeOptions: {
-							multipleValues: true,
-						},
-						options: [
-							{
-								name: 'headers',
-								displayName: 'Header',
-								values: [
-									{
-										displayName: 'Name',
-										name: 'name',
-										type: 'string',
-										required: true,
-										default: '',
-									},
-									{
-										displayName: 'Value',
-										name: 'value',
-										type: 'string',
-										default: '',
-									},
-								],
-							},
-						],
-						description: 'Custom headers to add to the email',
-					},
-					{
-						displayName: 'Reply To',
-						name: 'reply_to',
-						type: 'string',
-						default: '',
-						description: 'Reply-to email address. For multiple addresses, use comma-separated values.',
-					},
-					{
-						displayName: 'Scheduled At',
-						name: 'scheduled_at',
-						type: 'string',
-						default: '',
-						description: 'Schedule email to be sent later (e.g., "in 1 min" or ISO 8601 format)',
-					},
-					{
-						displayName: 'Tags',
-						name: 'tags',
-						type: 'fixedCollection',
-						default: { tags: [] },
-						typeOptions: {
-							multipleValues: true,
-						},
-						options: [
-							{
-								name: 'tags',
-								displayName: 'Tag',
-								values: [
-									{
-										displayName: 'Name',
-										name: 'name',
-										type: 'string',
-										required: true,
-										default: '',
-									},
-									{
-										displayName: 'Value',
-										name: 'value',
-										type: 'string',
-										required: true,
-										default: '',
-									},
-								],
-							},
-						],
-						description: 'Tags to attach to the email',
-					},
-					{
-						displayName: 'Topic ID',
-						name: 'topic_id',
-						type: 'string',
-						default: '',
-						description: 'Topic ID to scope the email to',
-					},
-				],
-			},
+			...emailOperations,
+			...templateOperations,
+			...domainOperations,
+			...apiKeyOperations,
+			...broadcastOperations,
+			...segmentOperations,
+			...topicOperations,
+			...contactOperations,
 
-			// EMAIL PROPERTIES - Send Batch operation
-			{
-				displayName: 'Emails',
-				name: 'emails',
-				type: 'fixedCollection',
-				required: true,
-				default: { emails: [{}] },
-				typeOptions: {
-					multipleValues: true,
-				},
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['sendBatch'],
-					},
-				},
-				description: 'Array of emails to send (max 100)',
-				options: [
-					{
-						name: 'emails',
-						displayName: 'Email',
-						values: [
-							{
-								displayName: 'Attachments',
-								name: 'attachments',
-								type: 'fixedCollection',
-								default: { attachments: [] },
-								typeOptions: {
-									multipleValues: true,
-								},
-								options: [
-									{
-										name: 'attachments',
-										displayName: 'Attachment',
-										values: [
-											{
-												displayName: 'Binary Property',
-												name: 'binaryPropertyName',
-												type: 'string',
-												default: 'data',
-												placeholder: 'data',
-												description: 'Name of the binary property which contains the file data',
-											},
-											{
-												displayName: 'Content ID',
-												name: 'content_id',
-												type: 'string',
-												default: '',
-												placeholder: 'image-1',
-												description: 'Content ID for embedding inline attachments via cid:',
-											},
-											{
-												displayName: 'Content Type',
-												name: 'content_type',
-												type: 'string',
-												default: '',
-												placeholder: 'application/pdf',
-												description: 'Content type for the attachment',
-											},
-											{
-												displayName: 'File Name',
-												name: 'filename',
-												type: 'string',
-												default: '',
-												placeholder: 'document.pdf',
-												description: 'Name for the attached file',
-											},
-										],
-									},
-								],
-								description: 'Attachments for this email (binary data only)',
-							},
-							{
-								displayName: 'BCC',
-								name: 'bcc',
-								type: 'string',
-								default: '',
-								description: 'BCC recipient email addresses (comma-separated)',
-							},
-							{
-								displayName: 'CC',
-								name: 'cc',
-								type: 'string',
-								default: '',
-								description: 'CC recipient email addresses (comma-separated)',
-							},
-							{
-								displayName: 'Content Type',
-								name: 'contentType',
-								type: 'options',
-								default: 'html',
-								options: [
-									{
-										name: 'HTML',
-										value: 'html',
-										description: 'Send email with HTML content',
-									},
-									{
-										name: 'HTML and Text',
-										value: 'both',
-										description: 'Send email with both HTML and text content',
-									},
-									{
-										name: 'Text',
-										value: 'text',
-										description: 'Send email with plain text content',
-									},
-									{
-										name: 'Template',
-										value: 'template',
-										description: 'Send email using a template',
-									},
-								],
-								description: 'Choose how to provide content for this email',
-							},
-							{
-								displayName: 'From',
-								name: 'from',
-								type: 'string',
-								required: true,
-								default: '',
-								placeholder: 'you@example.com',
-								description: 'Sender email address',
-							},
-							{
-								displayName: 'Headers',
-								name: 'headers',
-								type: 'fixedCollection',
-								default: { headers: [] },
-								typeOptions: {
-									multipleValues: true,
-								},
-								options: [
-									{
-										name: 'headers',
-										displayName: 'Header',
-										values: [
-											{
-												displayName: 'Name',
-												name: 'name',
-												type: 'string',
-												required: true,
-												default: '',
-											},
-											{
-												displayName: 'Value',
-												name: 'value',
-												type: 'string',
-												default: '',
-											},
-										],
-									},
-								],
-								description: 'Custom headers to add to the email',
-							},
-							{
-								displayName: 'HTML Content',
-								name: 'html',
-								type: 'string',
-								default: '',
-								description: 'HTML content of the email',
-								placeholder: '<p>Your HTML content here</p>',
-								typeOptions: {
-									rows: 4,
-								},
-								displayOptions: {
-									show: {
-										contentType: ['html', 'both'],
-									},
-								},
-							},
-							{
-								displayName: 'Reply To',
-								name: 'reply_to',
-								type: 'string',
-								default: '',
-								description: 'Reply-to email address. For multiple addresses, use comma-separated values.',
-							},
-							{
-								displayName: 'Subject',
-								name: 'subject',
-								type: 'string',
-								required: true,
-								default: '',
-								placeholder: 'Hello from n8n!',
-								description: 'Email subject',
-							},
-							{
-								displayName: 'Tags',
-								name: 'tags',
-								type: 'fixedCollection',
-								default: { tags: [] },
-								typeOptions: {
-									multipleValues: true,
-								},
-								options: [
-									{
-										name: 'tags',
-										displayName: 'Tag',
-										values: [
-											{
-												displayName: 'Name',
-												name: 'name',
-												type: 'string',
-												required: true,
-												default: '',
-											},
-											{
-												displayName: 'Value',
-												name: 'value',
-												type: 'string',
-												required: true,
-												default: '',
-											},
-										],
-									},
-								],
-								description: 'Tags to attach to the email',
-							},
-							{
-								displayName: 'Template Name or ID',
-								name: 'templateId',
-								type: 'options',
-								required: true,
-								default: '',
-								placeholder: '34a080c9-b17d-4187-ad80-5af20266e535',
-								typeOptions: {
-									loadOptionsMethod: 'getTemplates',
-								},
-								displayOptions: {
-									show: {
-										contentType: ['template'],
-									},
-								},
-								description: 'Select a template or enter an ID/alias using an expression. Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>.',
-							},
-							{
-								displayName: 'Template Variables',
-								name: 'templateVariables',
-								type: 'fixedCollection',
-								default: { variables: [] },
-								typeOptions: {
-									multipleValues: true,
-								},
-								displayOptions: {
-									show: {
-										contentType: ['template'],
-									},
-								},
-								description: 'Variables to render the template with',
-								options: [
-									{
-										name: 'variables',
-										displayName: 'Variable',
-										values: [
-											{
-												displayName: 'Key',
-												name: 'key',
-												type: 'options',
-												required: true,
-												default: '',
-												typeOptions: {
-													loadOptionsMethod: 'getTemplateVariables',
-													loadOptionsDependsOn: ['templateId'],
-													allowCustomValues: true,
-												},
-												description: 'Template variable name',
-											},
-											{
-												displayName: 'Value',
-												name: 'value',
-												type: 'string',
-												default: '',
-												description: 'Value for the template variable',
-											},
-										],
-									},
-								],
-							},
-							{
-								displayName: 'Text Content',
-								name: 'text',
-								type: 'string',
-								default: '',
-								description: 'Plain text content of the email',
-								typeOptions: {
-									rows: 4,
-								},
-								placeholder: 'Your plain text content here',
-								displayOptions: {
-									show: {
-										contentType: ['text', 'both'],
-									},
-								},
-							},
-							{
-								displayName: 'To',
-								name: 'to',
-								type: 'string',
-								required: true,
-								default: '',
-								placeholder: 'user@example.com',
-								description: 'Recipient email address (comma-separated for multiple)',
-							},
-							{
-								displayName: 'Topic ID',
-								name: 'topic_id',
-								type: 'string',
-								default: '',
-								description: 'Topic ID to scope the email to',
-							},
-						],
-					},
-				],
-			},
-			{
-				displayName: 'Batch Options',
-				name: 'batchOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['sendBatch'],
-					},
-				},
-				options: [
-					{
-						displayName: 'Idempotency Key',
-						name: 'idempotency_key',
-						type: 'string',
-						default: '',
-						description: 'Unique key to ensure the batch request is processed only once',
-					},
-					{
-						displayName: 'Validation Mode',
-						name: 'validation_mode',
-						type: 'options',
-						default: 'strict',
-						options: [
-							{
-								name: 'Strict',
-								value: 'strict',
-								description: 'Reject the entire batch if any email fails validation',
-							},
-							{
-								name: 'Permissive',
-								value: 'permissive',
-								description: 'Send valid emails and return errors for invalid ones',
-							},
-						],
-						description: 'How Resend should validate the batch',
-					},
-				],
-			},
-
-			// EMAIL PROPERTIES - Retrieve/Update/Cancel operations
-			{
-				displayName: 'Email ID',
-				name: 'emailId',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'ae2014de-c168-4c61-8267-70d2662a1ce1',
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['retrieve', 'update', 'cancel'],
-					},
-				},
-				description: 'The ID of the email to retrieve, update, or cancel',
-			},
-			{
-				displayName: 'Scheduled At',
-				name: 'scheduled_at',
-				type: 'string',
-				default: '',
-				placeholder: '2024-08-05T11:52:01.858Z',
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['update'],
-					},
-				},
-				description: 'Schedule email to be sent later. The date should be in ISO 8601 format (e.g., 2024-08-05T11:52:01.858Z).',
-			},
-			{
-				displayName: 'Return All',
-				name: 'returnAll',
-				type: 'boolean',
-				default: false,
-				displayOptions: {
-					show: {
-						resource: ['email', 'templates', 'domains', 'apiKeys', 'broadcasts', 'segments', 'topics', 'contacts'],
-						operation: ['list'],
-					},
-				},
-				description: 'Whether to return all results or only up to a given limit',
-			},
-			{
-				displayName: 'Limit',
-				name: 'limit',
-				type: 'number',
-				typeOptions: {
-					minValue: 1,
-				},
-				default: 50,
-				displayOptions: {
-					show: {
-						resource: ['email', 'templates', 'domains', 'apiKeys', 'broadcasts', 'segments', 'topics', 'contacts'],
-						operation: ['list'],
-						returnAll: [false],
-					},
-				},
-				description: 'Max number of results to return',
-			},
-			{
-				displayName: 'List Options',
-				name: 'emailListOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['email'],
-						operation: ['list'],
-					},
-				},
-				options: [
-					{
-						displayName: 'After',
-						name: 'after',
-						type: 'string',
-						default: '',
-						description: 'Return results after this email ID',
-					},
-					{
-						displayName: 'Before',
-						name: 'before',
-						type: 'string',
-						default: '',
-						description: 'Return results before this email ID',
-					},
-				],
-			},
-
-			// TEMPLATE PROPERTIES
-			{
-				displayName: 'Name',
-				name: 'templateName',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'order-confirmation',
-				displayOptions: {
-					show: {
-						resource: ['templates'],
-						operation: ['create'],
-					},
-				},
-				description: 'The name of the template',
-			},
-			{
-				displayName: 'From',
-				name: 'templateFrom',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'Resend Store <store@resend.com>',
-				displayOptions: {
-					show: {
-						resource: ['templates'],
-						operation: ['create'],
-					},
-				},
-				description: 'Sender email address. To include a friendly name, use the format "Your Name &lt;sender@domain.com&gt;".',
-			},
-			{
-				displayName: 'Subject',
-				name: 'templateSubject',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'Thanks for your order!',
-				displayOptions: {
-					show: {
-						resource: ['templates'],
-						operation: ['create'],
-					},
-				},
-				description: 'Default subject line for the template',
-			},
-			{
-				displayName: 'HTML Content',
-				name: 'templateHtml',
-				type: 'string',
-				required: true,
-				default: '',
-				typeOptions: {
-					multiline: true,
-					rows: 4,
-				},
-				placeholder: '<p>Name: {{{PRODUCT}}}</p><p>Total: {{{PRICE}}}</p>',
-				displayOptions: {
-					show: {
-						resource: ['templates'],
-						operation: ['create'],
-					},
-				},
-				description: 'HTML version of the template',
-			},
-			{
-				displayName: 'Template Variables',
-				name: 'templateVariables',
-				type: 'fixedCollection',
-				default: { variables: [] },
-				typeOptions: {
-					multipleValues: true,
-				},
-				displayOptions: {
-					show: {
-						resource: ['templates'],
-						operation: ['create', 'update'],
-					},
-				},
-				description: 'Define variables used in the template',
-				options: [
-					{
-						name: 'variables',
-						displayName: 'Variable',
-						values: [
-							{
-								displayName: 'Key',
-								name: 'key',
-								type: 'string',
-								required: true,
-								default: '',
-								description: 'Variable name (we recommend uppercase, e.g. PRODUCT)',
-							},
-							{
-								displayName: 'Type',
-								name: 'type',
-								type: 'options',
-								default: 'string',
-								options: [
-									{ name: 'String', value: 'string' },
-									{ name: 'Number', value: 'number' },
-								],
-								description: 'Variable data type',
-							},
-							{
-								displayName: 'Fallback Value',
-								name: 'fallbackValue',
-								type: 'string',
-								default: '',
-								description: 'Fallback value used when a variable is not provided',
-							},
-						],
-					},
-				],
-			},
-			{
-				displayName: 'Template Name or ID',
-				name: 'templateId',
-				type: 'options',
-				required: true,
-				default: '',
-				placeholder: '34a080c9-b17d-4187-ad80-5af20266e535',
-				typeOptions: {
-					loadOptionsMethod: 'getTemplates',
-				},
-				displayOptions: {
-					show: {
-						resource: ['templates'],
-						operation: ['get', 'update', 'delete'],
-					},
-				},
-				description: 'Select a template or enter an ID/alias using an expression. Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>.',
-			},
-			{
-				displayName: 'Update Fields',
-				name: 'templateUpdateFields',
-				type: 'collection',
-				placeholder: 'Add Field',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['templates'],
-						operation: ['update'],
-					},
-				},
-				options: [
-					{
-						displayName: 'Alias',
-						name: 'alias',
-						type: 'string',
-						default: '',
-						description: 'Template alias',
-					},
-					{
-						displayName: 'From',
-						name: 'from',
-						type: 'string',
-						default: '',
-						description: 'Sender email address',
-					},
-					{
-						displayName: 'HTML Content',
-						name: 'html',
-						type: 'string',
-						default: '',
-						typeOptions: {
-							multiline: true,
-							rows: 4,
-						},
-						description: 'HTML content of the template',
-					},
-					{
-						displayName: 'Name',
-						name: 'name',
-						type: 'string',
-						default: '',
-						description: 'Template name',
-					},
-					{
-						displayName: 'Reply To',
-						name: 'reply_to',
-						type: 'string',
-						default: '',
-						description: 'Reply-to email address. For multiple addresses, use comma-separated values.',
-					},
-					{
-						displayName: 'Subject',
-						name: 'subject',
-						type: 'string',
-						default: '',
-						description: 'Default subject line for the template',
-					},
-					{
-						displayName: 'Text Content',
-						name: 'text',
-						type: 'string',
-						default: '',
-						typeOptions: {
-							multiline: true,
-							rows: 4,
-						},
-						description: 'Plain text content. Set to an empty string to disable automatic plain text generation.',
-					},
-				],
-			},
-			{
-				displayName: 'List Options',
-				name: 'templateListOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['templates'],
-						operation: ['list'],
-					},
-				},
-				options: [
-					{
-						displayName: 'After',
-						name: 'after',
-						type: 'string',
-						default: '',
-						description: 'Return results after this template ID',
-					},
-					{
-						displayName: 'Before',
-						name: 'before',
-						type: 'string',
-						default: '',
-						description: 'Return results before this template ID',
-					},
-				],
-			},
-
-			// DOMAIN PROPERTIES
-			{
-				displayName: 'Domain Name',
-				name: 'domainName',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'example.com',
-				displayOptions: {
-					show: {
-						resource: ['domains'],
-						operation: ['create'],
-					},
-				},
-				description: 'The name of the domain you want to create',
-			},
-			{
-				displayName: 'Domain ID',
-				name: 'domainId',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: '4dd369bc-aa82-4ff3-97de-514ae3000ee0',
-				displayOptions: {
-					show: {
-						resource: ['domains'],
-						operation: ['get', 'verify', 'update', 'delete'],
-					},
-				},
-				description: 'The ID of the domain',
-			},
-			{
-				displayName: 'Additional Options',
-				name: 'additionalOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['domains'],
-						operation: ['create'],
-					},
-				},
-				options: [
-					{
-						displayName: 'Region',
-						name: 'region',
-						type: 'options',
-						options: [
-							{ name: 'US East 1', value: 'us-east-1' },
-							{ name: 'EU West 1', value: 'eu-west-1' },
-							{ name: 'South America East 1', value: 'sa-east-1' },
-							{ name: 'Asia Pacific Northeast 1', value: 'ap-northeast-1' },
-						],
-						default: 'us-east-1',
-						description: 'The region where emails will be sent from',
-					},
-					{
-						displayName: 'Custom Return Path',
-						name: 'custom_return_path',
-						type: 'string',
-						default: 'send',
-						description: 'Custom subdomain for the Return-Path address',
-					},
-				],
-			},
-			{
-				displayName: 'Domain Update Options',
-				name: 'domainUpdateOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['domains'],
-						operation: ['update'],
-					},
-				},
-				options: [
-					{
-						displayName: 'Click Tracking',
-						name: 'click_tracking',
-						type: 'boolean',
-						default: false,
-						description: 'Whether to track clicks within the body of each HTML email',
-					},
-					{
-						displayName: 'Open Tracking',
-						name: 'open_tracking',
-						type: 'boolean',
-						default: false,
-						description: 'Whether to track the open rate of each email',
-					},
-					{
-						displayName: 'TLS',
-						name: 'tls',
-						type: 'options',
-						options: [
-							{ name: 'Opportunistic', value: 'opportunistic' },
-							{ name: 'Enforced', value: 'enforced' },
-						],
-						default: 'opportunistic',
-						description: 'TLS setting for email delivery. Opportunistic attempts secure connection but falls back to unencrypted if needed. Enforced requires TLS and will not send if unavailable.',
-					},
-				],
-			},
-			{
-				displayName: 'List Options',
-				name: 'domainListOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['domains'],
-						operation: ['list'],
-					},
-				},
-				options: [
-					{
-						displayName: 'After',
-						name: 'after',
-						type: 'string',
-						default: '',
-						description: 'Return results after this domain ID',
-					},
-					{
-						displayName: 'Before',
-						name: 'before',
-						type: 'string',
-						default: '',
-						description: 'Return results before this domain ID',
-					},
-				],
-			},
-
-			// API KEY PROPERTIES
-			{
-				displayName: 'API Key Name',
-				name: 'apiKeyName',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'My API Key',
-				displayOptions: {
-					show: {
-						resource: ['apiKeys'],
-						operation: ['create'],
-					},
-				},
-				description: 'The name of the API key to create',
-			},
-			{
-				displayName: 'API Key ID',
-				name: 'apiKeyId',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'key_123456',
-				displayOptions: {
-					show: {
-						resource: ['apiKeys'],
-						operation: ['delete'],
-					},
-				},
-				description: 'The ID of the API key to delete',
-			},
-			{
-				displayName: 'Permission',
-				name: 'permission',
-				type: 'options',
-				options: [
-					{ name: 'Full Access', value: 'full_access' },
-					{ name: 'Sending Access', value: 'sending_access' },
-				],
-				default: 'full_access',
-				displayOptions: {
-					show: {
-						resource: ['apiKeys'],
-						operation: ['create'],
-					},
-				},
-				description: 'The permission level for the API key',
-			},
-			{
-				displayName: 'Domain ID',
-				name: 'domainId',
-				type: 'string',
-				default: '',
-				placeholder: '4dd369bc-aa82-4ff3-97de-514ae3000ee0',
-				displayOptions: {
-					show: {
-						resource: ['apiKeys'],
-						operation: ['create'],
-						permission: ['sending_access'],
-					},
-				},
-				description: 'Restrict an API key to send emails only from a specific domain. This is only used when the permission is set to sending access.',
-			},
-			{
-				displayName: 'List Options',
-				name: 'apiKeyListOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['apiKeys'],
-						operation: ['list'],
-					},
-				},
-				options: [
-					{
-						displayName: 'After',
-						name: 'after',
-						type: 'string',
-						default: '',
-						description: 'Return results after this API key ID',
-					},
-					{
-						displayName: 'Before',
-						name: 'before',
-						type: 'string',
-						default: '',
-						description: 'Return results before this API key ID',
-					},
-				],
-			},
-
-			// BROADCAST PROPERTIES
-			{
-				displayName: 'Broadcast ID',
-				name: 'broadcastId',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'bc_123456',
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-						operation: ['get', 'update', 'send', 'delete'],
-					},
-				},
-				description: 'The ID of the broadcast',
-			},
-			{
-				displayName: 'Segment ID',
-				name: 'segmentId',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'seg_123456',
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-						operation: ['create'],
-					},
-				},
-				description: 'The ID of the segment for this broadcast',
-			},
-			{
-				displayName: 'From',
-				name: 'broadcastFrom',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'you@example.com',
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-						operation: ['create'],
-					},
-				},
-				description: 'Sender email address. To include a friendly name, use the format "Your Name &lt;sender@domain.com&gt;".',
-			},
-			{
-				displayName: 'Subject',
-				name: 'broadcastSubject',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'Newsletter Subject',
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-						operation: ['create'],
-					},
-				},
-				description: 'Email subject',
-			},
-			{
-				displayName: 'HTML Content',
-				name: 'broadcastHtml',
-				type: 'string',
-				required: true,
-				default: '',
-				typeOptions: {
-					multiline: true,
-				},
-				placeholder: '<p>Your HTML content here with {{{FIRST_NAME|there}}} and {{{RESEND_UNSUBSCRIBE_URL}}}</p>',
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-						operation: ['create'],
-					},
-				},
-				description: 'The HTML version of the message. You can use variables like {{{FIRST_NAME|fallback}}} and {{{RESEND_UNSUBSCRIBE_URL}}}.',
-			},
-			{
-				displayName: 'Create Options',
-				name: 'broadcastCreateOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-						operation: ['create'],
-					},
-				},
-				options: [
-					{
-						displayName: 'Name',
-						name: 'name',
-						type: 'string',
-						default: '',
-						placeholder: 'Internal broadcast name',
-						description: 'The friendly name of the broadcast. Only used for internal reference.',
-					},
-					{
-						displayName: 'Reply To',
-						name: 'reply_to',
-						type: 'string',
-						default: '',
-						placeholder: 'noreply@example.com',
-						description: 'Reply-to email address. For multiple addresses, use comma-separated values.',
-					},
-					{
-						displayName: 'Text Content',
-						name: 'text',
-						type: 'string',
-						default: '',
-						typeOptions: {
-							multiline: true,
-						},
-						placeholder: 'Your plain text content here',
-						description: 'The plain text version of the message',
-					},
-					{
-						displayName: 'Topic ID',
-						name: 'topic_id',
-						type: 'string',
-						default: '',
-						placeholder: 'topic_123456',
-						description: 'Topic ID that the broadcast will be scoped to',
-					},
-				],
-			},
-			{
-				displayName: 'Update Fields',
-				name: 'broadcastUpdateFields',
-				type: 'collection',
-				placeholder: 'Add Field',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-						operation: ['update'],
-					},
-				},
-				options: [
-					{
-						displayName: 'From',
-						name: 'from',
-						type: 'string',
-						default: '',
-						placeholder: 'you@example.com',
-						description: 'Sender email address',
-					},
-					{
-						displayName: 'HTML Content',
-						name: 'html',
-						type: 'string',
-						default: '',
-						typeOptions: {
-							multiline: true,
-						},
-						placeholder: '<p>Your HTML content here</p>',
-						description: 'The HTML version of the message',
-					},
-					{
-						displayName: 'Name',
-						name: 'name',
-						type: 'string',
-						default: '',
-						placeholder: 'Internal broadcast name',
-						description: 'The friendly name of the broadcast. Only used for internal reference.',
-					},
-					{
-						displayName: 'Reply To',
-						name: 'reply_to',
-						type: 'string',
-						default: '',
-						placeholder: 'noreply@example.com',
-						description: 'Reply-to email address. For multiple addresses, use comma-separated values.',
-					},
-					{
-						displayName: 'Segment ID',
-						name: 'segment_id',
-						type: 'string',
-						default: '',
-						placeholder: 'seg_123456',
-						description: 'The ID of the segment you want to send to',
-					},
-					{
-						displayName: 'Subject',
-						name: 'subject',
-						type: 'string',
-						default: '',
-						placeholder: 'Newsletter Subject',
-						description: 'Email subject',
-					},
-					{
-						displayName: 'Text Content',
-						name: 'text',
-						type: 'string',
-						default: '',
-						typeOptions: {
-							multiline: true,
-						},
-						placeholder: 'Your plain text content here',
-						description: 'The plain text version of the message',
-					},
-					{
-						displayName: 'Topic ID',
-						name: 'topic_id',
-						type: 'string',
-						default: '',
-						placeholder: 'topic_123456',
-						description: 'Topic ID that the broadcast will be scoped to',
-					},
-				],
-			},
-			{
-				displayName: 'Send Options',
-				name: 'broadcastSendOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-						operation: ['send'],
-					},
-				},
-				options: [
-					{
-						displayName: 'Scheduled At',
-						name: 'scheduled_at',
-						type: 'string',
-						default: '',
-						placeholder: 'in 1 min',
-						description: 'Schedule the broadcast to be sent later (natural language or ISO 8601)',
-					},
-				],
-			},
-			{
-				displayName: 'List Options',
-				name: 'broadcastListOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['broadcasts'],
-						operation: ['list'],
-					},
-				},
-				options: [
-					{
-						displayName: 'After',
-						name: 'after',
-						type: 'string',
-						default: '',
-						description: 'Return results after this broadcast ID',
-					},
-					{
-						displayName: 'Before',
-						name: 'before',
-						type: 'string',
-						default: '',
-						description: 'Return results before this broadcast ID',
-					},
-				],
-			},
-
-			// SEGMENT PROPERTIES
-			{
-				displayName: 'Segment Name',
-				name: 'segmentName',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'Registered Users',
-				displayOptions: {
-					show: {
-						resource: ['segments'],
-						operation: ['create'],
-					},
-				},
-				description: 'The name of the segment',
-			},
-			{
-				displayName: 'Segment Name or ID',
-				name: 'segmentId',
-				type: 'options',
-				required: true,
-				default: '',
-				placeholder: 'seg_123456',
-				typeOptions: {
-					loadOptionsMethod: 'getSegments',
-				},
-				displayOptions: {
-					show: {
-						resource: ['segments'],
-						operation: ['get', 'delete'],
-					},
-				},
-				description: 'Select a segment or enter an ID using an expression. Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>.',
-			},
-			{
-				displayName: 'List Options',
-				name: 'segmentListOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['segments'],
-						operation: ['list'],
-					},
-				},
-				options: [
-					{
-						displayName: 'After',
-						name: 'after',
-						type: 'string',
-						default: '',
-						description: 'Return results after this segment ID',
-					},
-					{
-						displayName: 'Before',
-						name: 'before',
-						type: 'string',
-						default: '',
-						description: 'Return results before this segment ID',
-					},
-				],
-			},
-
-			// TOPIC PROPERTIES
-			{
-				displayName: 'Topic Name',
-				name: 'topicName',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'Weekly Newsletter',
-				displayOptions: {
-					show: {
-						resource: ['topics'],
-						operation: ['create'],
-					},
-				},
-				description: 'The name of the topic',
-			},
-			{
-				displayName: 'Default Subscription',
-				name: 'topicDefaultSubscription',
-				type: 'options',
-				required: true,
-				default: 'opt_in',
-				displayOptions: {
-					show: {
-						resource: ['topics'],
-						operation: ['create'],
-					},
-				},
-				options: [
-					{ name: 'Opt In', value: 'opt_in' },
-					{ name: 'Opt Out', value: 'opt_out' },
-				],
-				description: 'Default subscription preference for new contacts',
-			},
-			{
-				displayName: 'Create Options',
-				name: 'topicCreateOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['topics'],
-						operation: ['create'],
-					},
-				},
-				options: [
-					{
-						displayName: 'Description',
-						name: 'description',
-						type: 'string',
-						default: '',
-						description: 'Short description of the topic',
-					},
-					{
-						displayName: 'Visibility',
-						name: 'visibility',
-						type: 'options',
-						default: 'private',
-						options: [
-							{ name: 'Private', value: 'private' },
-							{ name: 'Public', value: 'public' },
-						],
-						description: 'Visibility on the unsubscribe page',
-					},
-				],
-			},
-			{
-				displayName: 'Topic Name or ID',
-				name: 'topicId',
-				type: 'options',
-				required: true,
-				default: '',
-				placeholder: 'topic_123456',
-				typeOptions: {
-					loadOptionsMethod: 'getTopics',
-				},
-				displayOptions: {
-					show: {
-						resource: ['topics'],
-						operation: ['get', 'update', 'delete'],
-					},
-				},
-				description: 'Select a topic or enter an ID using an expression. Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>.',
-			},
-			{
-				displayName: 'Update Fields',
-				name: 'topicUpdateFields',
-				type: 'collection',
-				placeholder: 'Add Field',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['topics'],
-						operation: ['update'],
-					},
-				},
-				options: [
-					{
-						displayName: 'Name',
-						name: 'name',
-						type: 'string',
-						default: '',
-						description: 'New name for the topic',
-					},
-					{
-						displayName: 'Description',
-						name: 'description',
-						type: 'string',
-						default: '',
-						description: 'New description for the topic',
-					},
-					{
-						displayName: 'Visibility',
-						name: 'visibility',
-						type: 'options',
-						default: 'private',
-						options: [
-							{ name: 'Private', value: 'private' },
-							{ name: 'Public', value: 'public' },
-						],
-						description: 'Visibility on the unsubscribe page',
-					},
-				],
-			},
-			{
-				displayName: 'List Options',
-				name: 'topicListOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['topics'],
-						operation: ['list'],
-					},
-				},
-				options: [
-					{
-						displayName: 'After',
-						name: 'after',
-						type: 'string',
-						default: '',
-						description: 'Return results after this topic ID',
-					},
-					{
-						displayName: 'Before',
-						name: 'before',
-						type: 'string',
-						default: '',
-						description: 'Return results before this topic ID',
-					},
-				],
-			},
-
-			// CONTACT PROPERTIES
-			{
-				displayName: 'Email',
-				name: 'email',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'contact@example.com',
-				displayOptions: {
-					show: {
-						resource: ['contacts'],
-						operation: ['create'],
-					},
-				},
-				description: 'The email address of the contact',
-			},
-			{
-				displayName: 'Contact Identifier',
-				name: 'contactIdentifier',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'e169aa45-1ecf-4183-9955-b1499d5701d3 or contact@example.com',
-				displayOptions: {
-					show: {
-						resource: ['contacts'],
-						operation: ['get', 'delete'],
-					},
-				},
-				description: 'The contact ID or email address',
-			},
-			{
-				displayName: 'Update By',
-				name: 'updateBy',
-				type: 'options',
-				options: [
-					{ name: 'Contact ID', value: 'id' },
-					{ name: 'Email Address', value: 'email' },
-				],
-				default: 'id',
-				displayOptions: {
-					show: {
-						resource: ['contacts'],
-						operation: ['update'],
-					},
-				},
-				description: 'Choose whether to update the contact by ID or email address',
-			},
-			{
-				displayName: 'Contact ID',
-				name: 'contactId',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'con_123456',
-				displayOptions: {
-					show: {
-						resource: ['contacts'],
-						operation: ['update'],
-						updateBy: ['id'],
-					},
-				},
-				description: 'The ID of the contact to update',
-			},
-			{
-				displayName: 'Contact Email',
-				name: 'contactEmail',
-				type: 'string',
-				required: true,
-				default: '',
-				placeholder: 'contact@example.com',
-				displayOptions: {
-					show: {
-						resource: ['contacts'],
-						operation: ['update'],
-						updateBy: ['email'],
-					},
-				},
-				description: 'The email address of the contact to update',
-			},
-			{
-				displayName: 'Create Fields',
-				name: 'contactCreateFields',
-				type: 'collection',
-				placeholder: 'Add Field',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['contacts'],
-						operation: ['create'],
-					},
-				},
-				options: [
-					{
-						displayName: 'First Name',
-						name: 'first_name',
-						type: 'string',
-						default: '',
-						description: 'The first name of the contact',
-					},
-					{
-						displayName: 'Last Name',
-						name: 'last_name',
-						type: 'string',
-						default: '',
-						description: 'The last name of the contact',
-					},
-					{
-						displayName: 'Properties',
-						name: 'properties',
-						type: 'fixedCollection',
-						default: { properties: [] },
-						typeOptions: {
-							multipleValues: true,
-						},
-						options: [
-							{
-								name: 'properties',
-								displayName: 'Property',
-								values: [
-									{
-										displayName: 'Key',
-										name: 'key',
-										type: 'string',
-										required: true,
-										default: '',
-									},
-									{
-										displayName: 'Value',
-										name: 'value',
-										type: 'string',
-										default: '',
-									},
-								],
-							},
-						],
-					},
-					{
-						displayName: 'Segments',
-						name: 'segments',
-						type: 'fixedCollection',
-						default: { segments: [] },
-						typeOptions: {
-							multipleValues: true,
-						},
-						options: [
-							{
-								name: 'segments',
-								displayName: 'Segment',
-								values: [
-									{
-										displayName: 'Segment ID',
-										name: 'id',
-										type: 'string',
-										required: true,
-										default: '',
-									},
-								],
-							},
-						],
-					},
-					{
-						displayName: 'Topics',
-						name: 'topics',
-						type: 'fixedCollection',
-						default: { topics: [] },
-						typeOptions: {
-							multipleValues: true,
-						},
-						options: [
-							{
-								name: 'topics',
-								displayName: 'Topic',
-								values: [
-									{
-										displayName: 'Topic ID',
-										name: 'id',
-										type: 'string',
-										required: true,
-										default: '',
-									},
-									{
-										displayName: 'Subscription',
-										name: 'subscription',
-										type: 'options',
-										default: 'opt_in',
-										options: [
-											{ name: 'Opt In', value: 'opt_in' },
-											{ name: 'Opt Out', value: 'opt_out' },
-										],
-									},
-								],
-							},
-						],
-					},
-					{
-						displayName: 'Unsubscribed',
-						name: 'unsubscribed',
-						type: 'boolean',
-						default: false,
-						description: 'Whether the contact is unsubscribed from emails',
-					},
-				],
-			},
-			{
-				displayName: 'Update Fields',
-				name: 'contactUpdateFields',
-				type: 'collection',
-				placeholder: 'Add Field',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['contacts'],
-						operation: ['update'],
-					},
-				},
-				options: [
-					{
-						displayName: 'First Name',
-						name: 'first_name',
-						type: 'string',
-						default: '',
-						description: 'The first name of the contact',
-					},
-					{
-						displayName: 'Last Name',
-						name: 'last_name',
-						type: 'string',
-						default: '',
-						description: 'The last name of the contact',
-					},
-					{
-						displayName: 'Unsubscribed',
-						name: 'unsubscribed',
-						type: 'boolean',
-						default: false,
-						description: 'Whether the contact is unsubscribed from emails',
-					},
-					{
-						displayName: 'Properties',
-						name: 'properties',
-						type: 'fixedCollection',
-						default: { properties: [] },
-						typeOptions: {
-							multipleValues: true,
-						},
-						options: [
-							{
-								name: 'properties',
-								displayName: 'Property',
-								values: [
-									{
-										displayName: 'Key',
-										name: 'key',
-										type: 'string',
-										required: true,
-										default: '',
-									},
-									{
-										displayName: 'Value',
-										name: 'value',
-										type: 'string',
-										default: '',
-									},
-								],
-							},
-						],
-					},
-				],
-			},
-			{
-				displayName: 'List Options',
-				name: 'contactListOptions',
-				type: 'collection',
-				placeholder: 'Add Option',
-				default: {},
-				displayOptions: {
-					show: {
-						resource: ['contacts'],
-						operation: ['list'],
-					},
-				},
-				options: [
-					{
-						displayName: 'After',
-						name: 'after',
-						type: 'string',
-						default: '',
-						description: 'Return results after this contact ID',
-					},
-					{
-						displayName: 'Before',
-						name: 'before',
-						type: 'string',
-						default: '',
-						description: 'Return results before this contact ID',
-					},
-				],
-			},
-
+			...emailFields,
+			...templateFields,
+			...domainFields,
+			...apiKeyFields,
+			...broadcastFields,
+			...segmentFields,
+			...topicFields,
+			...contactFields,
 		],
 	};
 	methods = {
 		loadOptions: {
-			async getTemplateVariables(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const getStringValue = (value: unknown) =>
-					typeof value === 'string' && value.trim() ? value : undefined;
-				const safeGet = (getter: () => unknown) => {
-					try {
-						return getter();
-					} catch {
-						return undefined;
-					}
-				};
-				const getParameterValue = (name: string) => {
-					const currentParameters = this.getCurrentNodeParameters();
-					const fromCurrentParameters = getStringValue(currentParameters?.[name]);
-					if (fromCurrentParameters) {
-						return fromCurrentParameters;
-					}
-
-					const fromCurrentNodeParameter = getStringValue(
-						safeGet(() => this.getCurrentNodeParameter(name)),
-					);
-					if (fromCurrentNodeParameter) {
-						return fromCurrentNodeParameter;
-					}
-
-					const fromNodeParameter = getStringValue(safeGet(() => this.getNodeParameter(name, '')));
-					if (fromNodeParameter) {
-						return fromNodeParameter;
-					}
-
-					return undefined;
-				};
-
-				const templateId = getParameterValue('emailTemplateId') ?? getParameterValue('templateId');
-				if (!templateId) {
-					return [];
-				}
-				const normalizedTemplateId = templateId.trim();
-				if (normalizedTemplateId.startsWith('={{') || normalizedTemplateId.includes('{{')) {
-					return [];
-				}
-
-				const credentials = await this.getCredentials('resendApi');
-				const apiKey = credentials.apiKey as string;
-
-				const response = await this.helpers.httpRequest({
-					url: `https://api.resend.com/templates/${encodeURIComponent(templateId)}`,
-					method: 'GET',
-					headers: {
-						Authorization: `Bearer ${apiKey}`,
-					},
-					json: true,
-				});
-
-				const variables = response?.variables ?? [];
-
-				return variables
-					.filter((variable: { key?: string }) => variable?.key)
-					.map((variable: { key: string; type?: string }) => {
-						const typeLabel = variable.type ? ` (${variable.type})` : '';
-						return {
-							name: `${variable.key}${typeLabel}`,
-							value: variable.key,
-						};
-					});
-			},
-			async getTemplates(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const credentials = await this.getCredentials('resendApi');
-				const apiKey = credentials.apiKey as string;
-				const returnData: INodePropertyOptions[] = [];
-				const limit = 100;
-				let after: string | undefined;
-				let hasMore = true;
-				let pageCount = 0;
-				const maxPages = 10;
-
-				while (hasMore) {
-					const qs: Record<string, string | number> = { limit };
-					if (after) {
-						qs.after = after;
-					}
-
-					const response = await this.helpers.httpRequest({
-						url: 'https://api.resend.com/templates',
-						method: 'GET',
-						headers: {
-							Authorization: `Bearer ${apiKey}`,
-						},
-						qs,
-						json: true,
-					});
-
-					const templates = response?.data ?? [];
-					for (const template of templates) {
-						if (!template?.id) {
-							continue;
-						}
-						const name = template.name ? `${template.name} (${template.id})` : template.id;
-						returnData.push({
-							name,
-							value: template.id,
-						});
-					}
-
-					hasMore = Boolean(response?.has_more);
-					after = templates.length ? templates[templates.length - 1].id : undefined;
-					pageCount += 1;
-					if (!after || pageCount >= maxPages) {
-						break;
-					}
-				}
-
-				return returnData;
-			},
-			async getSegments(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const credentials = await this.getCredentials('resendApi');
-				const apiKey = credentials.apiKey as string;
-				const returnData: INodePropertyOptions[] = [];
-				const limit = 100;
-				let after: string | undefined;
-				let hasMore = true;
-				let pageCount = 0;
-				const maxPages = 10;
-
-				while (hasMore) {
-					const qs: Record<string, string | number> = { limit };
-					if (after) {
-						qs.after = after;
-					}
-
-					const response = await this.helpers.httpRequest({
-						url: 'https://api.resend.com/segments',
-						method: 'GET',
-						headers: {
-							Authorization: `Bearer ${apiKey}`,
-						},
-						qs,
-						json: true,
-					});
-
-					const segments = response?.data ?? [];
-					for (const segment of segments) {
-						if (!segment?.id) {
-							continue;
-						}
-						const name = segment.name ? `${segment.name} (${segment.id})` : segment.id;
-						returnData.push({
-							name,
-							value: segment.id,
-						});
-					}
-
-					hasMore = Boolean(response?.has_more);
-					after = segments.length ? segments[segments.length - 1].id : undefined;
-					pageCount += 1;
-					if (!after || pageCount >= maxPages) {
-						break;
-					}
-				}
-
-				return returnData;
-			},
-			async getTopics(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const credentials = await this.getCredentials('resendApi');
-				const apiKey = credentials.apiKey as string;
-				const returnData: INodePropertyOptions[] = [];
-				const limit = 100;
-				let after: string | undefined;
-				let hasMore = true;
-				let pageCount = 0;
-				const maxPages = 10;
-
-				while (hasMore) {
-					const qs: Record<string, string | number> = { limit };
-					if (after) {
-						qs.after = after;
-					}
-
-					const response = await this.helpers.httpRequest({
-						url: 'https://api.resend.com/topics',
-						method: 'GET',
-						headers: {
-							Authorization: `Bearer ${apiKey}`,
-						},
-						qs,
-						json: true,
-					});
-
-					const topics = response?.data ?? [];
-					for (const topic of topics) {
-						if (!topic?.id) {
-							continue;
-						}
-						const name = topic.name ? `${topic.name} (${topic.id})` : topic.id;
-						returnData.push({
-							name,
-							value: topic.id,
-						});
-					}
-
-					hasMore = Boolean(response?.has_more);
-					after = topics.length ? topics[topics.length - 1].id : undefined;
-					pageCount += 1;
-					if (!after || pageCount >= maxPages) {
-						break;
-					}
-				}
-
-				return returnData;
-			},
+			getTemplateVariables,
+			getTemplates,
+			getSegments,
+			getTopics,
 		},
 	};
 	async execute(this: IExecuteFunctions): Promise<INodeExecutionData[][]> {
 		const items = this.getInputData();
 		const returnData: INodeExecutionData[] = [];
 		const length = items.length;
-		const normalizeEmailList = (value: string | string[] | undefined) => {
-			if (Array.isArray(value)) {
-				return value
-					.map((email) => String(email).trim())
-					.filter((email) => email);
-			}
-			if (typeof value === 'string') {
-				return value
-					.split(',')
-					.map((email) => email.trim())
-					.filter((email) => email);
-			}
-			return [];
-		};
-		const parseTemplateVariables = (
-			variablesInput: { variables?: Array<{ key: string; type: string; fallbackValue?: unknown }> } | undefined,
-			fallbackKey: 'fallbackValue' | 'fallback_value',
-			itemIndex: number,
-		) => {
-			if (!variablesInput?.variables?.length) {
-				return undefined;
-			}
-
-			return variablesInput.variables.map((variable) => {
-				const variableEntry: Record<string, unknown> = {
-					key: variable.key,
-					type: variable.type,
-				};
-
-				const fallbackValue = variable.fallbackValue;
-				if (fallbackValue !== undefined && fallbackValue !== '') {
-					let parsedFallback: string | number = fallbackValue as string;
-					if (variable.type === 'number') {
-						const numericFallback = typeof fallbackValue === 'number' ? fallbackValue : Number(fallbackValue);
-						if (Number.isNaN(numericFallback)) {
-							throw new NodeOperationError(
-								this.getNode(),
-								`Variable "${variable.key}" fallback value must be a number`,
-								{ itemIndex },
-							);
-						}
-						parsedFallback = numericFallback;
-					}
-					variableEntry[fallbackKey] = parsedFallback;
-				}
-
-				return variableEntry;
-			});
-		};
-		const buildTemplateSendVariables = (
-			variablesInput: { variables?: Array<{ key: string; value?: unknown }> } | undefined,
-		) => {
-			if (!variablesInput?.variables?.length) {
-				return undefined;
-			}
-			const variables: Record<string, unknown> = {};
-			for (const variable of variablesInput.variables) {
-				if (!variable.key) {
-					continue;
-				}
-				variables[variable.key] = variable.value ?? '';
-			}
-
-			return Object.keys(variables).length ? variables : undefined;
-		};
-		type ListOptions = {
-			after?: string;
-			before?: string;
-		};
-		const requestList = async (
-			url: string,
-			listOptions: ListOptions,
-			apiKey: string,
-			itemIndex: number,
-			returnAll: boolean,
-			limit?: number,
-		) => {
-			if (listOptions.after && listOptions.before) {
-				throw new NodeOperationError(
-					this.getNode(),
-					'You can only use either "After" or "Before", not both.',
-					{ itemIndex },
-				);
-			}
-
-			const shouldReturnAll = returnAll === true;
-			const qs: Record<string, string | number> = {};
-			const pageSize = shouldReturnAll ? 100 : (limit ?? 50);
-
-			if (pageSize !== undefined) {
-				qs.limit = pageSize;
-			}
-			if (listOptions.after) {
-				qs.after = listOptions.after;
-			}
-			if (listOptions.before) {
-				qs.before = listOptions.before;
-			}
-
-			const requestPage = () =>
-				this.helpers.httpRequest({
-					url,
-					method: 'GET',
-					headers: {
-						Authorization: `Bearer ${apiKey}`,
-					},
-					qs,
-					json: true,
-				});
-
-			if (!shouldReturnAll) {
-				const singleResponse = await requestPage();
-				if (
-					typeof limit === 'number' &&
-					limit > 0 &&
-					Array.isArray((singleResponse as { data?: unknown[] }).data)
-				) {
-					const responseData = (singleResponse as { data?: unknown[] }).data ?? [];
-					if (responseData.length > limit) {
-						(singleResponse as { data: unknown[] }).data = responseData.slice(0, limit);
-					}
-				}
-				return singleResponse;
-			}
-
-			const allItems: unknown[] = [];
-			let lastResponse: unknown;
-			let hasMore = true;
-			let pageCount = 0;
-			const maxPages = 100;
-			let paginationMode: 'after' | 'before' | undefined = listOptions.before ? 'before' : undefined;
-
-			while (hasMore) {
-				lastResponse = await requestPage();
-				const responseData = Array.isArray((lastResponse as { data?: unknown[] }).data)
-					? ((lastResponse as { data?: unknown[] }).data as unknown[])
-					: [];
-				allItems.push(...responseData);
-
-				hasMore = Boolean((lastResponse as { has_more?: boolean }).has_more);
-				pageCount += 1;
-				if (!hasMore || responseData.length === 0 || pageCount >= maxPages) {
-					break;
-				}
-
-				const lastItem = responseData[responseData.length - 1] as { id?: string } | undefined;
-				if (!lastItem?.id) {
-					break;
-				}
-
-				if (paginationMode === 'before') {
-					qs.before = lastItem.id;
-					delete qs.after;
-				} else {
-					qs.after = lastItem.id;
-					delete qs.before;
-					paginationMode = 'after';
-				}
-			}
-
-			if (lastResponse && Array.isArray((lastResponse as { data?: unknown[] }).data)) {
-				(lastResponse as { data: unknown[] }).data = allItems;
-				(lastResponse as { has_more?: boolean }).has_more = false;
-				return lastResponse;
-			}
-
-			return { object: 'list', data: allItems, has_more: false };
-		};
-
 		for (let i = 0; i < length; i++) {
 			try {
 				const resource = this.getNodeParameter('resource', i) as string;
@@ -3269,7 +490,7 @@ export class Resend implements INodeType {
 						const listOptions = this.getNodeParameter('emailListOptions', i, {}) as ListOptions;
 						const returnAll = this.getNodeParameter('returnAll', i) as boolean;
 						const limit = this.getNodeParameter('limit', i, 50) as number;
-						response = await requestList('https://api.resend.com/emails', listOptions, apiKey, i, returnAll, limit);
+						response = await requestList(this, 'https://api.resend.com/emails', listOptions, apiKey, i, returnAll, limit);
 
 					} else if (operation === 'retrieve') {
 						const emailId = this.getNodeParameter('emailId', i) as string;
@@ -3331,7 +552,7 @@ export class Resend implements INodeType {
 							html: templateHtml,
 						};
 
-						const variables = parseTemplateVariables(templateVariables, 'fallbackValue', i);
+						const variables = parseTemplateVariables(this, templateVariables, 'fallbackValue', i);
 						if (variables?.length) {
 							requestBody.variables = variables;
 						}
@@ -3362,6 +583,7 @@ export class Resend implements INodeType {
 						const returnAll = this.getNodeParameter('returnAll', i) as boolean;
 						const limit = this.getNodeParameter('limit', i, 50) as number;
 						response = await requestList(
+							this,
 							'https://api.resend.com/templates',
 							listOptions,
 							apiKey,
@@ -3400,7 +622,7 @@ export class Resend implements INodeType {
 							requestBody.text = updateFields.text;
 						}
 
-						const variables = parseTemplateVariables(templateVariables, 'fallbackValue', i);
+						const variables = parseTemplateVariables(this, templateVariables, 'fallbackValue', i);
 						if (variables?.length) {
 							requestBody.variables = variables;
 						}
@@ -3498,7 +720,7 @@ export class Resend implements INodeType {
 						const listOptions = this.getNodeParameter('domainListOptions', i, {}) as ListOptions;
 						const returnAll = this.getNodeParameter('returnAll', i) as boolean;
 						const limit = this.getNodeParameter('limit', i, 50) as number;
-						response = await requestList('https://api.resend.com/domains', listOptions, apiKey, i, returnAll, limit);
+						response = await requestList(this, 'https://api.resend.com/domains', listOptions, apiKey, i, returnAll, limit);
 
 					} else if (operation === 'delete') {
 						const domainId = this.getNodeParameter('domainId', i) as string;
@@ -3544,7 +766,7 @@ export class Resend implements INodeType {
 						const listOptions = this.getNodeParameter('apiKeyListOptions', i, {}) as ListOptions;
 						const returnAll = this.getNodeParameter('returnAll', i) as boolean;
 						const limit = this.getNodeParameter('limit', i, 50) as number;
-						response = await requestList('https://api.resend.com/api-keys', listOptions, apiKey, i, returnAll, limit);
+						response = await requestList(this, 'https://api.resend.com/api-keys', listOptions, apiKey, i, returnAll, limit);
 
 					} else if (operation === 'delete') {
 						const apiKeyId = this.getNodeParameter('apiKeyId', i) as string;
@@ -3684,6 +906,7 @@ export class Resend implements INodeType {
 						const returnAll = this.getNodeParameter('returnAll', i) as boolean;
 						const limit = this.getNodeParameter('limit', i, 50) as number;
 						response = await requestList(
+							this,
 							'https://api.resend.com/broadcasts',
 							listOptions,
 							apiKey,
@@ -3739,7 +962,7 @@ export class Resend implements INodeType {
 						const listOptions = this.getNodeParameter('segmentListOptions', i, {}) as ListOptions;
 						const returnAll = this.getNodeParameter('returnAll', i) as boolean;
 						const limit = this.getNodeParameter('limit', i, 50) as number;
-						response = await requestList('https://api.resend.com/segments', listOptions, apiKey, i, returnAll, limit);
+						response = await requestList(this, 'https://api.resend.com/segments', listOptions, apiKey, i, returnAll, limit);
 
 					} else if (operation === 'delete') {
 						const segmentId = this.getNodeParameter('segmentId', i) as string;
@@ -3798,7 +1021,7 @@ export class Resend implements INodeType {
 						const listOptions = this.getNodeParameter('topicListOptions', i, {}) as ListOptions;
 						const returnAll = this.getNodeParameter('returnAll', i) as boolean;
 						const limit = this.getNodeParameter('limit', i, 50) as number;
-						response = await requestList('https://api.resend.com/topics', listOptions, apiKey, i, returnAll, limit);
+						response = await requestList(this, 'https://api.resend.com/topics', listOptions, apiKey, i, returnAll, limit);
 
 					} else if (operation === 'update') {
 						const topicId = this.getNodeParameter('topicId', i) as string;
@@ -3942,7 +1165,7 @@ export class Resend implements INodeType {
 						const listOptions = this.getNodeParameter('contactListOptions', i, {}) as ListOptions;
 						const returnAll = this.getNodeParameter('returnAll', i) as boolean;
 						const limit = this.getNodeParameter('limit', i, 50) as number;
-						response = await requestList('https://api.resend.com/contacts', listOptions, apiKey, i, returnAll, limit);
+						response = await requestList(this, 'https://api.resend.com/contacts', listOptions, apiKey, i, returnAll, limit);
 
 					} else if (operation === 'delete') {
 						const contactIdentifier = this.getNodeParameter('contactIdentifier', i) as string;
