+++ nodes/Resend/descriptions/index.ts
@@ -1,0 +1,8 @@
+export * from './EmailDescription';
+export * from './TemplateDescription';
+export * from './DomainDescription';
+export * from './ApiKeyDescription';
+export * from './BroadcastDescription';
+export * from './SegmentDescription';
+export * from './TopicDescription';
+export * from './ContactDescription';
